Tiny main.go refactor
Okay now I'm done poking at it. The next job is to configure the animator based on a CLI arg, so that it can load other apps as controlled by React. The plan is to host these on separate pages, since there isn't much of a use-case for hot-swapping between them