Finish moving the rubik's cube to the new system
This means that the rubik's cube shader now sits underneath the 2 extra
virtual canvases that enable zooming and panning (despite the fact that
it doesn't make use of this). I've made it this way because it allows me
to reuse a lot more webgl setup code than I otherwise could.

A little annoyingly, this extra jump between framebuffers does impact
performance a little, and in particular it affects the output resolution
for the rubik's cube demo a bit.